(perf): Drop sample-first allocators; trait-only sizing (Constant unified)
The protocol's 1D + ND batch allocating callables, and Constant's 1D + ND
one-shot batch wrappers, called the scalar/persistent kernel for the first
query just to read `typeof(first_val)` for the output buffer. That extra
evaluation is per-batch overhead on a hot path — measurable at small N and
strictly redundant once the trait predicts correctly.

`_output_eltype(itp, Tq)` (and the 3-arg `_output_eltype(Tv, Tg, Tq)` used
by one-shot wrappers) now drives all allocators directly: empty-batch and
populated paths share one expression. The kernel still produces the same
runtime values; only the container eltype is sized from types, not from
sampling.

Constant follows the same natural-promotion rule as every other method (no
Constant-specific helper or override per the original Issue #144 PR's
direction). The kernel's `* one(dL)` still preserves scalar-path semantics
(Int×Int×Int scalar callable returns Int via kernel direct); batch and
one-shot wrappers widen via the trait's Int→Float upgrade. Tests pinning
the old sample-first `Vector{Int}` for fully-Int Constant batch are
updated to `Vector{Float64}`.

Locations:
- src/core/interpolant_protocol.jl:60-79 (1D batch)
- src/core/interpolant_protocol.jl:255-272 (ND batch)
- src/constant/constant_oneshot.jl:277-298 (1D one-shot batch)
- src/constant/nd/constant_nd_oneshot.jl:98-115 (ND one-shot batch)

+++ test/test_constant.jl
@@ -195,10 +195,10 @@ end
         @test out ≈ [10.0, 20.0, 30.0]
 
         # Test 3: Vector allocating with Integer grid + Integer query.
-        # Constant duck-types output to `eltype(y)` — no Float widening.
+        # Natural promotion: trait upgrades Int→Float at allocation.
         result_vec = constant_interp(x_int, y_int, [0, 1, 2])
-        @test result_vec isa Vector{Int}
-        @test result_vec == [10, 20, 30]
+        @test result_vec isa Vector{Float64}
+        @test result_vec == [10.0, 20.0, 30.0]
 
         # Test 4: In-place Real→Float wrapper (lines 440-468)
         # constant_interp!(output, x::AbstractVector{T}, y::AbstractVector{T}, x_targets::AbstractVector{S})
